+++ packages/core/src/tools/urlPolyfill.spec.ts
@@ -1,4 +1,4 @@
-import { isFirefox } from './specHelper'
+import { isFirefox } from '../../test/specHelper'
 import { getHash, getOrigin, getPathName, getSearch, isValidUrl, normalizeUrl } from './urlPolyfill'
 import { getLocationOrigin } from './utils'
