+++ packages/rum-core/src/domain/requestCollection.spec.ts
@@ -1,17 +1,19 @@
 import {
   Configuration,
   DEFAULT_CONFIGURATION,
-  FetchStub,
-  FetchStubManager,
-  isIE,
   RequestType,
   resetFetchProxy,
   resetXhrProxy,
+} from '@datadog/browser-core'
+import {
+  FetchStub,
+  FetchStubManager,
+  isIE,
   SPEC_ENDPOINTS,
   stubFetch,
   stubXhr,
   withXhr,
-} from '@datadog/browser-core'
+} from '../../../core/test/specHelper'
 import { LifeCycle, LifeCycleEventType } from './lifeCycle'
 import { RequestCompleteEvent, RequestStartEvent, trackFetch, trackXhr } from './requestCollection'
 import { clearTracingIfCancelled, Tracer } from './tracing/tracer'
