+++ packages/rum-core/src/domain/rumSession.spec.ts
@@ -3,11 +3,11 @@ import {
   COOKIE_ACCESS_DELAY,
   DEFAULT_CONFIGURATION,
   getCookie,
-  isIE,
   SESSION_COOKIE_NAME,
   setCookie,
   stopSessionManagement,
 } from '@datadog/browser-core'
+import { isIE } from '../../../core/test/specHelper'
 
 import { LifeCycle, LifeCycleEventType } from './lifeCycle'
 import { RUM_SESSION_KEY, RumTrackingType, startRumSession } from './rumSession'
