+++ packages/rum-core/src/domain/tracing/tracer.spec.ts
@@ -1,4 +1,5 @@
-import { Configuration, DEFAULT_CONFIGURATION, isIE, objectEntries } from '@datadog/browser-core'
+import { Configuration, DEFAULT_CONFIGURATION, objectEntries } from '@datadog/browser-core'
+import { isIE } from '../../../../core/test/specHelper'
 import { setup, TestSetupBuilder } from '../../../test/specHelper'
 import {
   RumFetchCompleteContext,
