+++ packages/rum-recorder/src/domain/privacy.spec.ts
@@ -1,4 +1,4 @@
-import { isIE } from '@datadog/browser-core'
+import { isIE } from '../../../core/test/specHelper'
 import {
   nodeShouldBeHidden,
   nodeOrAncestorsShouldBeHidden,
