+++ packages/rum-recorder/src/domain/rrweb/mutationObserver.spec.ts
@@ -1,4 +1,4 @@
-import { isIE } from '@datadog/browser-core'
+import { isIE } from '../../../../core/test/specHelper'
 import { createMutationPayloadValidator } from '../../../test/utils'
 import { snapshot, NodeType } from '../rrweb-snapshot'
 import { MutationController } from './mutation'
